Updating

+++ SWAS/Resource.cpp
@@ -135,7 +135,7 @@ void Resource::SetResourceCount(double resourceCount)
 	_resourceCount = resourceCount;
 }
 
-int Resource::GetResourceCount()
+double Resource::GetResourceCount()
 {
 	return _resourceCount;
 }
